Add extra write task
